# vanlife-frontend/docs/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "OpenAPI definition",
      "version": "v0"
    },
    "servers": [
      {
        "url": "http://localhost:9000",
        "description": "Generated server url"
      }
    ],
    "paths": {
...
      "/api/locations": {
        "get": {
          "tags": [
            "location-controller"
          ],
          "operationId": "getAllLocations",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Location"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "location-controller"
          ],
          "operationId": "createLocation",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/Location"
                  }
                }
              }
            }
          }
        }
      },
      "/api/locations/{locationId}/tags": {
        "post": {
          "tags": [
            "location-controller"
          ],
          "operationId": "associateTagsWithLocation",
          "parameters": [
            {
              "name": "locationId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "uniqueItems": true,
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/Location"
                  }
                }
              }
            }
          }
        }
      },
      "/api/comments": {
        "post": {
          "tags": [
            "comment-controller"
          ],
          "operationId": "addComment",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        }
      },
      "/api/auth/google": {
        "post": {
          "tags": [
            "google-auth-controller"
          ],
          "operationId": "handleGoogleAuth",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleCode"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/ratings/location/{locationId}": {
        "get": {
          "tags": [
            "rating-controller"
          ],
          "operationId": "getRatingsByLocation",
          "parameters": [
            {
              "name": "locationId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Rating"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/locations/{id}": {
        "get": {
          "tags": [
            "location-controller"
          ],
          "operationId": "getLocationById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/Location"
                  }
                }
              }
            }
          }
        }
      },
      "/api/locations/search": {
        "get": {
          "tags": [
            "location-controller"
          ],
          "operationId": "findLocationsByTag",
          "parameters": [
            {
              "name": "tagName",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Location"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/comments/location/{locationId}": {
        "get": {
          "tags": [
            "comment-controller"
          ],
          "operationId": "getCommentsByLocation",
          "parameters": [
            {
              "name": "locationId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Comment"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/locations/{locationId}/tags/{tagId}": {
        "delete": {
          "tags": [
            "location-controller"
          ],
          "operationId": "removeTagFromLocation",
          "parameters": [
            {
              "name": "locationId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            },
            {
              "name": "tagId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/Location"
                  }
                }
              }
            }
          }
        }
      },
      "/api/comments/{id}": {
        "delete": {
          "tags": [
            "comment-controller"
          ],
          "operationId": "deleteComment",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string"
            },
            "picture": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "roles": {
              "type": "string"
            },
            "active": {
              "type": "boolean"
            }
          }
        },
        "Tag": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "name": {
              "type": "string"
            },
            "predefined": {
              "type": "boolean"
            },
            "description": {
              "type": "string"
            }
          }
        },
...
        "Comment": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "text": {
              "type": "string"
            },
            "user": {
              "$ref": "#/components/schemas/User"
            },
            "location": {
              "$ref": "#/components/schemas/Location"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "Location": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "latitude": {
              "type": "number",
              "format": "double"
            },
            "longitude": {
              "type": "number",
              "format": "double"
            },
            "address": {
              "type": "string"
            },
            "mapServiceId": {
              "type": "string"
            },
            "mapUrl": {
              "type": "string"
            },
            "ratings": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Rating"
              }
            },
            "comments": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Comment"
              }
            },
            "createdBy": {
              "$ref": "#/components/schemas/User"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "updatedAt": {
              "type": "string",
              "format": "date-time"
            },
            "amenities": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "photoUrls": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "safetyRating": {
              "type": "number",
              "format": "double"
            },
            "accessibility": {
              "type": "string"
            },
            "cost": {
              "type": "string"
            },
            "popularity": {
              "type": "integer",
              "format": "int32"
            },
            "tags": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Tag"
              }
            }
          }
        },
        "Rating": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "score": {
              "type": "integer",
              "format": "int32"
            },
            "user": {
              "$ref": "#/components/schemas/User"
            },
            "location": {
              "$ref": "#/components/schemas/Location"
            }
          }
        },
...
        "GoogleCode": {
          "type": "object",
          "properties": {
            "code": {
              "type": "string"
            }
          }
        }
      }
    }
  }